+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New features:
 
 * `-Xbacktraces.raise` will print Ruby backtraces whenever an exception is
   raised.
+* `Java.import { import }` blocks for importing Java classes.
 
 Changes: